pyproject.toml
[project]
name = "creato"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.1.0",
    "requests>=2.32.4",
    "streamlit>=1.45.1",
]
